tests/test_requirements.txt
# Additional requirements for DeepSeek testing
# These are in addition to the main requirements.txt

# Azure AI Inference SDK for testing DeepSeek alternative approach
azure-ai-inference>=1.0.0

# Ensure we have the latest LiteLLM for Azure compatibility
litellm>=1.65.0

# For better error handling and debugging
requests>=2.31.0
